refactor: move from one monolithic file into multiple modular files for easier usage and maintainability
