+++ crates/sentinel/src/collection/mod.rs
@@ -1,0 +1,16 @@
+pub mod aggregation;
+pub mod collection;
+pub mod operations;
+pub mod query;
+pub mod streaming;
+pub mod tests;
+pub mod verification;
+pub mod wal;
+
+pub use collection::*;
+pub use operations::*;
+pub use streaming::*;
+pub use query::*;
+pub use verification::*;
+pub use aggregation::*;
+pub use wal::*;
